requirements/base.txt
influxdb==5.0.0
flask==1.0.2
PyYAML==3.12
munch==2.3.2

